refactor: hanum pay qr screen scanner refactor
