+++ src/components/hanumPay/index.ts
@@ -1,1 +1,2 @@
 export * from './header';
+export * from './QRScanner';
