+++ src/screens/hanumPay/qr/index.tsx
@@ -1,45 +1,16 @@
-import { SafeAreaView, View } from 'react-native';
 import React from 'react';
-import QRCodeScanner from 'react-native-qrcode-scanner';
-import { BarCodeReadEvent } from 'react-native-camera';
 
-import { HanumPayHeader, Text } from 'src/components';
-import { colors } from 'src/styles';
+import { HanumPayHeader, QRScanner } from 'src/components';
 
 import * as S from './styled';
 
 export const HanumPayQRScreen: React.FC = () => {
-  /** 바코드가 감지되면 실행되는 함수 */
-  const onSuccess = (e: BarCodeReadEvent) => {
-    console.log(e, 'data');
-  };
   return (
     <S.HanumPayQRWrapper>
-       <S.HanumPayQRHeaderWrapper>
-         <HanumPayHeader title="결제하기" />
+      <S.HanumPayQRHeaderWrapper>
+        <HanumPayHeader title="결제하기" />
       </S.HanumPayQRHeaderWrapper>
-      <SafeAreaView>
-    <QRCodeScanner
-        onRead={onSuccess}
-        showMarker={true}
-        // vibration change
-        reactivate={true}
-        reactivateTimeout={0}
-        fadeIn={false}
-        customMarker={
-                <S.HanumPayQRBoxContainer>
-                  <Text.Column>
-                    <Text size={16} isCenter color={colors.white}>
-                      한움페이 결제 QR코드를 {'\n'}아래 상자에 맞춰주세요
-                    </Text>
-                  </Text.Column>
-                  <S.HanumPayQRBox />
-                </S.HanumPayQRBoxContainer>
-              }
-      />
-      </SafeAreaView>
-      </S.HanumPayQRWrapper>
+      <QRScanner />
+    </S.HanumPayQRWrapper>
   );
 };
-
-
